Agent invocation cost rounded every row to zero
`agent_invocations.cost_cents` was declared integer and the writer did
Math.round(). A model call in this system costs a FRACTION of a cent —
the measured average across 2,118 gateway completions is 0.045, and the
cheapest agent's entire per-output budget is 0.2 — so every value became
0. All 1,587 rows recorded zero cost, and the ledger reported with total
confidence that the agent fleet had cost nothing.

That is the exact failure this console was built to stop, committed in
the ledger written to prevent it. Caught by an adversarial verification
pass over the map of this area, not by any test — nothing asserted the
column could hold what it was given.

Now numeric(12,6), matching events.cost_cents, which never lost it and is
where the console's spend figures come from. The rows written before the
migration are all zero and cannot be recovered — the precision was
destroyed at write time — so the column carries a COMMENT saying not to
read cost from them. They are kept rather than deleted because the table
is append-only evidence.

+++ packages/db/src/migrations/0024_agent_invocations.sql
@@ -33,7 +33,9 @@ CREATE TABLE IF NOT EXISTS agent_invocations (
   -- businesses, customers, leads, sessions and builds.
   subject_id         text,
   trace_id           text,
-  cost_cents         integer NOT NULL DEFAULT 0,
+  -- Fractional cents. A model call here costs well under one cent, so an
+  -- integer column rounds every row to zero (see migration 0025).
+  cost_cents         numeric(12,6) NOT NULL DEFAULT 0,
   -- ⛔ The success monitor. False means the escalation ladder was used, which
   -- costs more and means the champion did not produce a valid output first try.
   first_pass         boolean NOT NULL,
